Add/delete parties - needs err/validations

+++ src/app/home/action.ts
@@ -2,14 +2,16 @@
 import { createClient } from "@/utils/supabase/server";
 import { Party } from "./page";
 import { revalidatePath } from "next/cache";
-import { redirect } from "next/navigation";
+import { AuthError, Session, User, UserResponse } from "@supabase/supabase-js";
 
-export async function deleteParties(parties: Party[]) {
+export async function deleteParties(parties: Party[]): Promise<void> {
   "use server"
   const supabase = await createClient();
   for (const party of parties) {
     await supabase.from("party").delete().eq("id", party.id);
-    await supabase.from("build").delete().eq("id", party.leader);
+    if(party.leader) {
+      await supabase.from("build").delete().eq("id", party.leader);
+    }
     if (party.second) {
       await supabase.from("build").delete().eq("id", party.second);
     }
@@ -18,6 +20,24 @@ export async function deleteParties(parties: Party[]) {
     }
   }
   revalidatePath("/home");
-  //needs a better page refresh
-
 }
+
+
+export async function addParty(args: {name:string, game: string}): Promise<void> {
+  "use server"
+  const {name, game} = args;
+  const supabase = await createClient();
+  const user_id = await (await supabase.auth.getUser()).data.user?.id;
+  console.log(user_id);
+  const {error} = await supabase.from("party").insert({ //row level security what..
+    description: "",
+    name,
+    game,
+    user_id
+  })
+  if(error) {
+    console.log("insert err: "+error.message)
+  }
+  revalidatePath("/home");
+}
+
